+++ joulescope_ui/recording_viewer_device_v2.py
@@ -439,13 +439,13 @@ class RecordingViewerDeviceV2:
         for signal in reader.signals.values():
             dual_markers = {}
 
-            def cbk(timestamp, annotation_type, data):
+            def cbk(timestamp, annotation_type, group_id, data):
                 if signal.signal_type == SignalType.FSR:
                     # convert to seconds
                     # todo make this respect UTC, when UTC is implemented
                     timestamp = timestamp / signal.sample_rate
                 if annotation_type == AnnotationType.TEXT:
-                    self._cmdp.invoke('!Widgets/Waveform/annotation/add', [signal.name, timestamp, data])
+                    self._cmdp.invoke('!Widgets/Waveform/annotation/add', [signal.name, timestamp, group_id, data])
                 elif annotation_type == AnnotationType.MARKER:
                     if data[-1] in 'ab':
                         name = data[:-1]