+++ CHANGELOG.md
@@ -6,10 +6,12 @@ This file contains the list of changes made to pyjoulescope_ui.
 
 ## 0.9.5
 
-2021 Mar 4 [in progress]
+2021 Mar 6 [in progress]
 
 *   Fixed delta time computation (dual markers) for JLS v2.
 *   Fixed crash on viewing JLS v2 signal with no data.
+*   Fixed min/max fill color.
+*   Added annotation group_id support.
 
 
 ## 0.9.4